genai-platform/backend/ml/requirements.txt
torch
Pillow
transformers
python-dotenv
google-generativeai
transformers==4.38.2
torch>=2.0.0
Pillow>=9.0.0
python-docx>=0.8.11
PyMuPDF==1.23.3
google-generativeai==0.3.2
python-dotenv>=1.0.0
transformers==4.38.2
torch>=2.0.0
Pillow>=9.0.0
google-generativeai==0.3.2
python-dotenv>=1.0.0







